Split parser from converter
